# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Resident Management Service API
  description: Canonical target-state API documentation for Resident Management Service in Project A (Group 1 - Apartment Identity and Resident Management).
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: http://localhost:8000
    description: API Gateway route
paths:
...
  /api/v1/owners:
    get:
      tags:
        - Owners
      summary: "OWN-001: List owner profiles"
      description: Retrieve paginated owner profiles. Restricted to SYSTEM_ADMINISTRATOR and APARTMENT_MANAGER.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, ARCHIVED]
        - name: search
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Owners retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResidentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      tags:
        - Owners
      summary: "OWN-002: Create owner profile"
      description: Creates a new owner profile. Validates user existence with identity-access-service.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
      responses:
        '201':
          description: Owner profile created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleResidentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Owner profile already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '503':
          description: Identity service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/owners/{ownerId}:
    get:
      tags:
        - Owners
      summary: "OWN-003: Get owner profile"
      description: Retrieve a specific owner profile. Allowed for SYSTEM_ADMINISTRATOR, APARTMENT_MANAGER, or the owner themselves.
      security:
        - BearerAuth: []
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Owner retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleResidentResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Owner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT token issued/resigned by API Gateway. User JWT for public endpoints, Service JWT for internal endpoints.
  schemas:
    ProfileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, ARCHIVED]
    CreateProfileRequest:
      type: object
      required:
        - userId
        - firstName
        - lastName
        - email
      properties:
        userId:
          type: string
          format: uuid
          description: Identity Access user UUID
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
...
    PagedData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProfileResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    SingleResidentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          $ref: '#/components/schemas/ProfileResponse'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    PagedResidentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          $ref: '#/components/schemas/PagedData'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        details:
          nullable: true
    ApiErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid